requirements.txt
pbr>=0.5.21,<1.0
amqplib>=0.6.1
anyjson>=0.3.3
argparse
Babel>=1.3
eventlet>=0.13.0
greenlet>=0.3.2
iso8601>=0.1.8
kombu>=2.4.8
lockfile>=0.8
lxml>=2.3
netaddr
oslo.config>=1.2.0
paramiko>=1.8.0
Paste
PasteDeploy>=1.5.0
python-glanceclient>=0.9.0
python-keystoneclient>=0.3.2
python-novaclient>=2.15.0
python-swiftclient>=1.5
requests>=1.1
Routes>=1.12.3
six>=1.4.1
SQLAlchemy>=0.7.8,<=0.7.99
sqlalchemy-migrate>=0.7.2
stevedore>=0.10
suds>=0.4
WebOb>=1.2.3,<1.3
wsgiref>=0.1.2
